app/requirements.txt
flask==2.2.2
flask_cors==3.0.10
psycopg2-binary==2.9.5
torch==1.13.1
pandas==1.5.2
apscheduler==3.9.1
httpx
fastapi
uvicorn[standard]
openpyxl
pymsteams